ref #667: Correctly resolve mlt field when adding records
This fixes the resolution of the MLT field name when adding a new record
through an 'embeded' table editor. Before this it was assumed that the
field on the source table was always named `{TARGET_TABLE}_mlt` - if a
custom name along with `connectedTableName` was used, then adding a new
record through the 'add' button of embeded table editors would fail.