match cms_tpl_page only on non-empty pagePath
fixes #713

With the update to Symfony 4, the behavior of the router changed
slightly in regards to how it handles trailing slashes in urls. See:
https://symfony.com/doc/4.1/routing.html#redirecting-urls-with-trailing-slashes
This resulted in the fact that now the home route does not match if it
has a trailing slash. (/en/)
That is because the cms_tpl_page route also matched on trailing slash.

The fix is to stop cms_tpl_page to match on an empty pagePath.