ref #63427: improve documentation of obscure mapping in core.
The Mapping needed to be introduced because the directories used for packages in the mono repo where changed when we introduced the mono repo. Undoing the name change is too much effort and works in the wrong direction. Fixing the calls to the \TViewPathManager::getObjectPackageViewPath - which would be the correct way to go about this - is currently out of scope as well. So for now we introduce a mapping - with the aim to remove it once we rework / get rid of the old \TViewPathManager::getObjectPackageViewPath